Miscellaneous cleanup
1. We had an undocumented / untested feature where you can override the
HTTP method via a header. This changeset removes that. CXF has a
slightly different header for doing exactly the same thing. Not sure
about RESTEasy.

2. We had a whole lot of unnecessary override methods in
FHIRHttpServletRequestWrapper. Since the parent class defaults to
calling the delegate already, its cleaner to leave those out of this
subclass.

3. Committing minor edits to the JAX-RS Search resource...it doesn't
actually consume APPLICATION_FHIR media types in any of the methods and
there's no need to override the class-level path with the same value

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>