#3084 - CQL error when running with us-core installed
In the base parameter definitions, the parameter name and parameter code
values match. The parameter resolver code used the name for comparisons
and lookup and that functions just fine. The us-core parameter
definition name and code do not match up and the name value isn't useful
from a comparison and lookup point of view. I switched everything over
to use the code value and that works better both with and without
us-core installed.

Signed-off-by: Corey Sanders <corey.thecolonel@gmail.com>