Fix CQL Engine model resolution for 4.3.0
In order to make CQL easier to write against a FHIR data model, the CQL
translator automatically inserts some CQL functions from the FHIRHelpers
library into generated ELM. Most commonly used, are the ToString
overloads that take in named code objects (e.g. AdministrativeGender)
and automatically call `.value`, so that CQL can be treat the code value
as a String.

Those overloads of ToString must all be resolved the first time the
function is used so that the engine can pick the correct overload to use
for the code object in question. This includes looking at overloads that
take in ExposureState and EvidenceVariableType code objects that were
removed in 4.3.0. Prior to this changeset the `FhirModelResolver` would
not be able to resolve the arguments for those overloads into actual
model objects and it caused an NPE to be generated. I've updated the
logic to treat those missing named code objects as generic code objects
and that fixes the problem. The rest of this change is just reenabling
tests that were disabled until this problem was fixed.

Signed-off-by: Corey Sanders <corey.thecolonel@gmail.com>