#3061 - Fix NPE when Measure refers to non-existent library (#3187)
* #3061 - Fix NPE when Measure refers to non-existent library

I applied a fix before the holiday that corrected a category of NPE
errors related to execution against a measure or library resources that
don't exist. It turns out that wasn't the core of the issue 3061. This
goes back and addresses when a valid measure reference is used, but that
Measure refers to a Library resource that does not exist. I also added
some handling that allows resolution of Library resources by reference
in addition to by canonical URL in the registry and also updated the
error handling around retrieving the primary library reference from the
Measure resource.

Signed-off-by: Corey Sanders <corey.thecolonel@gmail.com>

* Update copyright dates

Signed-off-by: Corey Sanders <corey.thecolonel@gmail.com>

* Whitespace corrections and added javadocs

Signed-off-by: Corey Sanders <corey.thecolonel@gmail.com>