#3195 - Bubble up exception message from CQL engine
The CQL operations were wrapping the exception from the CQL engine in a
FHIROperationException as assuming that the nested exception would
automatically be included in the issues list which isn't the case at
all. I added logic to manually add the details from the CQLException to
the issues list so that it shows up in the OperationOutcome.

Signed-off-by: Corey Sanders <corey.thecolonel@gmail.com>