issue #3289 - rationalize ResourceType and ResourceTypeCode (#3290)
* issue #3289 - ResourceType and ResourceTypeCode

1. Update the CodeGenerator to combine the ResourceType and
FHIRResourceType code subtypes into a single new code subtype named
`ResourceTypeCode`
2. Rename the ResourceTypeName enum in fhir-core to ResourceType, move
the `from(String)` method there, and update ResourceTypeCode to use this
enum instead of duplicating the values in a nested class
3. Update all uses of ResourceType.Value and FHIRResourceType.Value to
use the new ResourceType enum instead

The result is
1. a simple enum we can use for ResourceType constants throughout the
codebase; and
2. a single code subtype that builds on this enum for when its used
within the fhir model

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>

* Favor ResourceType over ResourceTypeCode when not in the model

Standardize on ResourceType constants instead of ResourceTypeCode
constants.
The main difference is that ResourceTypeCode extends Code (which is an
Element) and so it can have an optional id and extensions.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>

* fixup copyrights

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>

* update copyright on the rest of the generated classes

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>