issue #3232- add the extra IN clause more selectively
Previously, we were adding this extra IN clause for the _type parameter
in all cases.
In the case of a whole-system search *count* query, this was slowing
things down a bit. We still add the IN clause to the query where it
seemed to help, but now we avoid that for the count query.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>