issues #3216 and #2512 - bundle ref resolution from FHIRPathResolve
In addition, this commit adds support for the `resource:` scheme in our
Reference type checking / reference value validation. This scheme is
used in the SMART health cards spec and so I think it makes sense to
explicitly allow it.

I thought that I would need to do additional logic for supporting
reference rewriting for this scheme but, based on the test I added to
FHIRRestHelperTest, apparently not...we just needed to let it past our
parser.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>