issue #2512 - allow any URI scheme in literal reference values
Per discussion at
https://chat.fhir.org/#narrow/stream/179166-implementers/topic/Allowed.20uri.20schemes.20for.20Bundle.2Eentry.2EfullUrl.20and.20referenc.2E.2E.2E/near/271636029

We'll still reject literal reference values that begin or end with `:`
and anything that looks like a "relative" URL path but doesn't follow
the FHIR REST URL Pattern (e.g. a literal value like "123" with no
scheme prefix).

I also provided some algorithm examples in a comment and added a single
local Bundle reference resolution test for the Bundle-in-a-Bundle case.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>