Attempt to compute search parameter config if missing in ParametersUtil
Currently, if a tenant gets added / fixed it requires a server restart
to take effect.
With these changes, we'll log a warning and then attempt to compute the
missing tenant search parameter config. Its still computed only once,
but this should enable better support for dynamically adding new
tenants.

Additionally, ParametersSearchUtilTest and
MultiTenantSearchParameterTest now use an 'extended' tenant to test
search-parameter-extensions (instead of 'default') so-as to avoid
conflicts with other tests.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>