issue #3385 - ParametersUtil and CompartmentUtil init changes
I made the constructors public and now we initialize the underlying maps
on construction instead of during static initialization.

This is done under-the-covers when you call SearchUtil.init() and,
therefor, calling SearchUtil.init() is no longer optional.
Initially I planned to tackle SearchUtil refactoring as well, but that
was more than I wanted to bite off at this point in time.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>