complete-mock update to avoid accidental override of Patient compartment
I finally got to the bottom of the strange test failure I was seeing.
1. We have serverRegistryResourceProviderEnabled
2. Conditionally, we load the database with all the sample data from
fhir-examples
3. Now the maps in CompartmentUtil are member variables and they get
populated when the instance is constructed
4. Our EraseOperation implementation instantiates its own
CompartmentUtil for checking whether the target of $erase is in the
Patient compartment or not
5. The EraseOperationTest fires AFTER the spec/example
tests, so this instance of CompartmentUtil was loading the Patient
CompartmentDefinition uploaded from fhir-examples in step number 2.

To work around the issue, I updated our CompleteMockDataGenerator to
omit the CompartmentDefinition.code CompartmentType and provide only an extension
instead. I also had to add CompartmentDefinition cleanup for our migration
tests since the previous/last releases had the CompartmentDefinition
instance that causes the trouble.

I also updated the CompleteMockDataGenerator to skip Period.start
elements to avoid violating the `per-1` constraint.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>