issue #3091 - Add resources to their own compartment
This works for Patient, Encounter, Practitioner, and RelatedPerson
compartments.
For Device, we still extract the membership parameter, but `GET
Device/{id}/Device` is considered an invalid search.

This addresses the issue outlined in #3091 because the `fhir-smart`
module is turning our searches into compartment searches under the
covers. Now the Patient resource that corresponds to the in-context
patient will be returned from that.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>