issue #3447 - add expanded valuesets to fhir-core-r4b
I thought that we needed the hl7.terminology package so that we could
expand the R4B ValueSets, but it turns out the R4B downloads page has
its own "expansions.json" download that contains all of the necessary
ValueSets in their expanded form.

Now only fhir-core-r4b will be packaged on the server by default...if
someone needs the other 3000+ resources from hl7.terminology they can
still add that to the userlib.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>