Add two more v3 CodeSystems that are used in tests
I presume these were in the tests because they existed in some of the
data and so, instead of changing the expectation of the tests, I opted
to add these v3 codesystems to the fhir-core-r4b ResourceProvider
instead.

Note: in one case
(`SearchTest.testSearchConditionEvidenceCaseSensitivity`) the expected
behavior didn't make sense to me and so, even after adding the
CodeSystem, I changed the expectation of the test.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>