fixups for Group export and EverythingOperation
For Group/$export, this changeset restores a line I had accidentally deleted
which sets the Group.id for Group export.

For Patient/$everything, now that we're properly reading the fhir-server-config,
we need to adjust the tenant1 configuration so that all resource types
that actually get used are in the config so that we can set `open` to false.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>