handle delete entries in history response from fhir-smart
This fixes an issue that we introduced with #3423 where we reject
history interactions that include one or more DELETE entries in their
response bundle (because we cannot check them for patient compartment
membership since they contain no resource content).

However, now that we reject history interactions that involve patient
compartment resource types (unless the approving scope is of type 'user'
or 'system'), it should always be safe to ignore the delete entries in
the afterHistory call.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>