issue #3507 - treat "undelete" updates like create-on-update in changes
Update db2 and postgresql add_any_resource.sql stored procs (and
DerbyResourceDAO.storeResource) to write a 'C' instead of a 'U' for this
case. This way, system and type-level history interactions can properly
report the response status as a 201 instead of a 200.

Note, however, that there are ongoing
limitations for the accuracy of instance-level history entries as
outlined at
https://github.com/IBM/FHIR/issues/3507#issuecomment-1081157116

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>