Defensive code in FHIRPersistenceJDBCImpl.changes()
We found that, due to the issue with ResourceTypeName, a system export
was invoking FHIRPersistenceJDBCImpl.changes() with a list of type names
that includes invalid values.

Previously, that led to an invalid SQL statement.  Now we will filter
out the invalid names (and throw if some were passed but none were
found).

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>