issue #3592 - use compartment membership for patient resource type
Previously, for scopes granting patient-scoped Patient write access
(patient/Patient.write), the beforeRead, beforeVread, and beforeHistory
methods prevented any interaction with any Patient resource unless that
resource had one of the ids passed in the `patient_id` claim of the
access token.

Now we attempt to perform those interactions and rely on the compartment
membership check in the afterRead, afterVread, and afterHistory methods
to ensure that this is part of the in-context patient(s)' record.
To avoid leaking "does patient id xyz exist" info, we reject reading of
non-existent patient resources with a 403 (unlike the usual 404)...this
also prevents 'create-on-update' behavior for this resource type.

To accommodate the change, FHIRRestHelper.doRead is updated so that
`afterRead` is invoked prior to throwing the
FHIRPersistenceResourceNotFoundException that becomes a 404.

Unfortunately, because `afterRead` is never called for a read of a
deleted resource, a user can now determine if any given resource id has
a current status of DELETED (but shouldn't be able to do much with that
info). To change that, we'd need to stop throwing from the PL for this
case.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>