issue #3619 - support removal of primitive values via FHIRPath
The FHIRPathPatch API is implemented via utilities in FHIRPathUtil.
Both the FHIRPatchPatch API and the unerlying utility methods were
designed to operate on FHIR nodes (FHIRPathResourceNode and
FHIRPathElementNode).

However, the delete API in particular is not strictly bound to paths
that select FHIR nodes. Previously, calling delete on a path that
selects one of these nodes would result in a NullPointerException.
With the previous commit, it would instead become a no-op.
With this commit, we will now attempt to remove the primitive value node
and rebuild the parent accordingly.

This allows users to remove fields like Resource.id and Element.id via
FHIRPath and it even allows you to remove the value of primitive FHIR
elements (e.g. Patient.active.value) in cases where those primitives
already have an extension (otherwise they need to have a value).

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>