issue #3641 - pre-expand select valuesets via the VSAC Term Service
Added a fledgling VSACRegistryResourceProvider and used it to expand US
Core ValueSets which were composed from VSAC value sets.

The expansion step was added to the ResourceProcessor.

Additionally, we were failing to validate the QuestionnaireResponse examples
due to a new constraint that uses an unsupported FHIRPath function.
I added a workaround and opened #3646 to get proper support added.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>