issue #3654 - update ConstraintGenerator to strip version suffix
ConstraintGenerator.getProfiles() was called from 4 places in this
class.
All but one of those was for building the argument to the FHIRPath
`extension()` function.
That means that there should be no version suffix and so I renamed the
method to `getProfilesWithoutVersion` and updated it accordingly.

In the single other spot (`generateProfileConstraint(Node)`), I added
the `getProfile()` logic in-line.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>