Edge case fix for matching codeSetMap entries
Fixes an edge case where if a valueset expansion has values from
two different versions of the same codesystem and the code being looked
up specifies no version and exists in the second one. Previously, we
just returned the result of the 'contains' check on the first set of
codes.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>