Fix FHIRVersionRequestFilter and `_type` query parameter handling
A single test case was failing. I fixed it by finding and addressing a
few separate issues:

1. In FHIRVersionRequestFilter, we were looking for the fhirVersion from
the Content-Type header for POST interactions; this was failing for POST
interactions to the `_search` endpoint and so now we check the Accept
header for that case instead.

2. Even after this fix, it was failing when the client set an explicit
`_type` parameter. This goes back to
https://github.com/IBM/FHIR/issues/3012 which is about some
preprocessing being done by CXF on the Accept header. Now that Liberty
has a fix for that, I removed
`org.apache.cxf.jaxrs.mediaTypeCheck.strict` form our web.xml and set
`jaxrs.cxf.use.noop.requestPreprocessor=true` in bootstrap.properties
instead. This has knock-on effects for clients...forcing them to use
proper MediaTypes on their requests. For example, requests to the
$bulkdata-status endpoint will no longer be allowed to use
application/fhir+json (only `application/json`).

3. Our main Dockerfile was overwriting the default bootstrap.properties
file and so I removed that as well. The other option would be to add
this noop.requestPreprocessor to the dockerfile's bootstrap.properties
as well.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>