issue #3665 - handle coding in valueset when we can't validate display (#3669)
* issue #3665 - handle coding in valueset but can't validate display

Previously, FHIRTermServer.validateCode(ValueSet, CodeableConcept)
returned a false ValidationOutcome in cases where the code was a valid
member of the ValueSet but includes a display value that we couldn't
validate (e.g. due to missing the corresponding CodeSystem).

Now, we will return true in the ValidationOutcome for this case, but we
include a message alongside it to indicate that we didn't actually
validate the display value. From the FHIRPath MemberOfFunction, we
translate that into an informational OperationOutcomeIssue and add it to
the EvaluationContext.

For example:
> Unable to validate display value 'Diastolic Blood Pressure' for code
'8462-4' in system 'http://loinc.org'

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>

* code formatting

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>

Co-authored-by: Robin Arnold <robin.arnold23@ibm.com>

Co-authored-by: Robin Arnold <robin.arnold23@ibm.com>