issue #3448 - convert property-based concept hiearchies in getConcepts
1. added `hasPropertyHierarchy` and `convertToSimpleCodeSystem`
utilities to CodeSystemSupport. the conversion utility uses JGraphT to
simplify the creation of an acyclic directed graph and then traverses
that structure to build an updated CodeSystem resource with nested
concepts. this allows the rest of our CodeSystemSupport to stay the
same.

2. call `convertToSimpleCodeSystem` from
RegistryTermServiceProvider.getConcepts to convert property-based
hierarchy into the more-common nested concept flavor so that we can use
our existing CodeSystemSupport filter approach for CodeSystems in the
registry

3. added hand-crafted CodeSystem resources that test parent-to-child and
child-to-parent edges with the conversion logic

4. added tests to verify our updated expansions match the expansions
from tx.fhir.org for select ValueSets that draw from polyhierarchical
CodeSystems like v3-ActCode and v3-RoleCode

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>