issue #3735 - remove invoke(String,String,Resource) special processing
For some reason, our exception handling for
`invoke(String,String,Resource)` differs from the exception handling in
all the other invoke methods.

I assume that was done so that the $invoke operation properly returns a
200 OK in cases where the data is invalid.
But this is not the right level for that logic and it can prevent us
from seeing important server issues in the logs.
I double-checked and the ValidateOperation is
already returning a valid OperationOutcome response for successful
validation of invalid resources, so I think we just need to remove the
special handling at the Operation JAX-RS resource.

Additionally, we found "interesting" liberty behavior that was
preventing us from seeing the exception stack trace even when FINE
logging is enabled; that log message is now level INFO so that the stack
trace gets printed (but it will still only be printed when the FINE
level is enabled).

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>