issue #3503 - special case underlying $everything "Forbidden" exceptions
In addition to the change, I added a new e2e test package under
fhir-server-test to confirm the fix. The test will conditionally execute
based on whether the server advertises its
.well-known/smart-configuration (which we only serve when its
configured). There's a ton of cases we could add, but for now I just
wanted to cover the specific case in #3503. Previously, we had only unit
tests for the fhir-smart interceptor.

I also removed an unused private method from the fhir-smart interceptor.

Finally, I snuck in a couple very minor updates:
* removed an unused test-jar dependency from fhir-core-r4b
* a single word change to a specific AuthzPolicyEnforcement message to
hint that only "SMART" scopes are echoed back to the client (and not
other scopes that may have been passed)

Signed-off-by: Lee Surprenant <lmsurpre@merative.com>