issue #3655 - throw from ServerFHIRRetrieveProvider for failed reads
In my opinion, neither the old behavior (500 server error) nor the new
before (200 OK with a null expression result) are ideal.
However, since we didn't intend to change it, I've restored the previous
behavior for now.
The most relevant advantage is we'll now get a reasonable error message
when the user tries to evaluate expressions on resources that don't
exist.

Signed-off-by: Lee Surprenant <lmsurpre@merative.com>