Make fhir-persistence-blob-app more robust
This one hasn't seen much usage yet and I hit a couple minor issues.
1. I had accidentally set `--fhir-config-dir` to defaultServer/config
when its supposed to be to defaultServer. I expected a good error
message but instead it just failed.
2. The docs claim `--tenant-id` and `--ds-id` are option (defaulting to
"default") but when I left them out of the `--create-container` action I
got a NullPointerExceptioon

These changes fix the NPE by initializing tenantId and dsId to "default"
and add an extra check to provide a better error message when a bad
directory is passed in `--fhir-config-dir`.

Signed-off-by: Lee Surprenant <lmsurpre@merative.com>