issue #3874 - use legacy export when payload offloading is configured
A cleaner approach would be to add a method to FHIRPersistence so that
implementations can indicate whether they support fetchResourcePayloads
which is what we use for the fast export. Then we could pass that into
BulkDataClient.submitExport where we decide which export implementation
to use.

However, ideally we'd just add this support...then we probably wouldn't
need those changes.

Therefor, for now, I opted for the quick and dirty fix.

Signed-off-by: Lee Surprenant <lmsurpre@merative.com>