issue #3845 - update release workflow to push to ghcr.io
Previously we injected Docker Hub credentials, logged in using those,
and published our images to Docker Hub under ibmcom.

Now the release workflow will use the docker/login-action to login to
the the GitHub Container Registry (ghcr.io) with the injected
GITHUB_TOKEN and then push there instead of Docker Hub.

Consequently, 1_dockerhub.sh was renamed to 1_docker.sh
and we added the registry prefix to the dockerfile-maven-plugin
repository config.

Signed-off-by: Lee Surprenant <lmsurpre@merative.com>