issue #3859 - track retired tables in the model so we can drop them
And use that information to skip migration steps on these resource types
as well.

After the previous commit for #3859, dropping a schema that was created
with a 4.x version of fhir-persistence-schema was failing because we
never dropped the tables that have been removed from the model.
With these changes, we now flag the removed resource types as "retired"
and avoid their creation in new schemas that way (vs. leaving them out
of the model entirely).

We could potentially remove this logic if we ever get to
https://github.com/LinuxForHealth/FHIR/issues/3263

Signed-off-by: Lee Surprenant <lmsurpre@merative.com>