Throw when an invalid resource type is passed
Previously we attempted to create the table even when the resource type
was invalid.
This got a little weird now that we avoid creation of the tables when
the resource type has been retired.

With these changes, now a retired resource type will result in a warning
(but no tables), and an invalid resource type will result in a failure
(IllegalArgumentException).

Signed-off-by: Lee Surprenant <lmsurpre@merative.com>