issue #3919 - validate Resource.id when invoked at instance level
I also did some major refactoring so that we can return 200 OK for more
cases.

Specifically, I tried to split the types of validation into two:
* If preconditions of the $validate operation are not met, we continue
to throw and thus return HTTP 400 Bad Request responses.
* If preconditions pass, but the interaction being validated (via
'mode') is not supported for the passed input, we now reutnr HTTP 200 OK
with an OperationOutcome that contains one or more issues with severity
ERROR.

Signed-off-by: Lee Surprenant <lmsurpre@merative.com>