issue #3958 - replace AES encryption with Hashids
The encryption was used to obscure the incrementing job id and not as a
secure cryptographic function.
However, it was flagged in CodeQL as insecure and the use of
"encryption" in the code made it less clear that repeatable
encryption/decryption was expected/desired (and not a vulnerability).

Here I replaced the encryption with the use of a separate library called
"Hashids" which is fit-for-purpose without implying imaginary security;
its just a convenient way to convert between numeric values and short,
unique string values for use in URLs.

Signed-off-by: Lee Surprenant <lmsurpre@merative.com>