Replace JobIdEncodingTransformerTest with static utils
Now that we can encode and decode the job with minimal effort, the
caching done by JobIdEncodingTransformer was overkill.

Additionally, because the encoded ids are simpler, I don't think we need
to support raw job ids any more.  Thus, instead of defaulting to null
(i.e. no encoding), the new default will be a string of "" and thus all
job ids will always be encoded.

Signed-off-by: Lee Surprenant <lmsurpre@merative.com>