Another `capabilities` change
The HL7 validator issues a warning on our current CapabilityStatement
because `name` is intended for computer processing (so the recommend
against whitespace but still allow it).

`title` is the field intended for human consumption and so I think its
cleaner to use that instead.

I also removed the no-longer-needed calls to the static `string()`
factory...because thats what I do.

Signed-off-by: Lee Surprenant <lmsurpre@merative.com>